fix: Modified brandId and productId to Long Modified brandId and productId to Long, as it was in Entity class

+++ src/main/java/com/jamesjss/retail/prices/application/repository/PricesRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 //Output Port
 public interface PricesRepository {
-    List<Prices> searchByDateProductAndBrand(LocalDateTime dateBetween, Integer productId, Integer brandId);
+    List<Prices> searchByDateProductAndBrand(LocalDateTime dateBetween, Long productId, Long brandId);
 }
