+++ src/main/java/com/jamesjss/retail/prices/infrastructure/adapter/PricesJpaRepository.java
@@ -7,5 +7,5 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public interface PricesJpaRepository extends JpaRepository<PricesEntity, Long> {
-    List<PricesEntity> findByBrandIdAndProductIdAndStartDateLessThanEqualAndEndDateGreaterThanEqual(Integer brandId, Integer productId, LocalDateTime startDate, LocalDateTime endTime);
+    List<PricesEntity> findByBrandIdAndProductIdAndStartDateLessThanEqualAndEndDateGreaterThanEqual(Long brandId, Long productId, LocalDateTime startDate, LocalDateTime endTime);
 }
